# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/plans/permissions": {
      "post": {
        "operationId": "PlansController_createPermission",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermissionDefinitionDto"
              },
              "examples": {
                "shipments": {
                  "summary": "Permiso de gestión de envíos",
                  "value": {
                    "code": "shipments.manage",
                    "description": "Permite crear, editar y eliminar envíos",
                    "feature": "shipments"
                  }
                },
                "drivers": {
                  "summary": "Permiso de gestión de conductores",
                  "value": {
                    "code": "drivers.manage",
                    "description": "Permite gestionar conductores",
                    "feature": "drivers"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Permiso creado exitosamente"
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "description": "Unauthorized — JWT inválido o ausente"
          },
          "403": {
            "description": "Sin permisos"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear un nuevo permiso",
        "tags": ["Plans"]
      },
      "get": {
        "operationId": "PlansController_findAllPermissions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de permisos"
          },
          "401": {
            "description": "Unauthorized — JWT inválido o ausente"
          },
          "403": {
            "description": "Sin permisos"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar todos los permisos",
        "tags": ["Plans"]
      }
    },
    "/api/v1/plans/permissions/{id}": {
      "get": {
        "operationId": "PlansController_findOnePermission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permiso encontrado"
          },
          "401": {
            "description": "Unauthorized — JWT inválido o ausente"
          },
          "403": {
            "description": "Forbidden — sin permisos suficientes"
          },
          "404": {
            "description": "Permiso no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener un permiso por ID",
        "tags": ["Plans"]
      },
      "patch": {
        "operationId": "PlansController_updatePermission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePermissionDefinitionDto"
              },
              "examples": {
                "update": {
                  "summary": "Actualizar descripción",
                  "value": {
                    "description": "Permite gestionar todos los envíos de la empresa"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Permiso actualizado"
          },
          "401": {
            "description": "Unauthorized — JWT inválido o ausente"
          },
          "403": {
            "description": "Sin permisos"
          },
          "404": {
            "description": "Permiso no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar un permiso por ID",
        "tags": ["Plans"]
      },
      "delete": {
        "operationId": "PlansController_removePermission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permiso eliminado"
          },
          "401": {
            "description": "Unauthorized — JWT inválido o ausente"
          },
          "403": {
            "description": "Sin permisos"
          },
          "404": {
            "description": "Permiso no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar un permiso por ID",
        "tags": ["Plans"]
      }
    },
    "/api/v1/plans/assign-permission": {
      "post": {
        "operationId": "PlansController_assignPermission",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignPermissionDto"
              },
              "examples": {
                "assign": {
                  "summary": "Asignar permiso de envíos al plan Starter",
                  "value": {
                    "planId": "uuid-del-plan",
                    "permissionId": "uuid-del-permiso"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Permiso asignado al plan"
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "description": "Unauthorized — JWT inválido o ausente"
          },
          "403": {
            "description": "Sin permisos"
          },
          "404": {
            "description": "Plan o permiso no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Asignar un permiso a un plan",
        "tags": ["Plans"]
      }
    },
    "/api/v1/plans": {
      "post": {
        "operationId": "PlansController_createPlan",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanDto"
              },
              "examples": {
                "basic": {
                  "summary": "Plan Básico",
                  "value": {
                    "name": "Starter",
                    "description": "Plan para pequeñas empresas de transporte",
                    "price": 19.99,
                    "interval": "month",
                    "is_active": true
                  }
                },
                "premium": {
                  "summary": "Plan Premium",
                  "value": {
                    "name": "Business",
                    "description": "Plan para empresas con flota mediana",
                    "price": 49.99,
                    "interval": "month",
                    "is_active": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plan creado exitosamente"
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "description": "Unauthorized — JWT inválido o ausente"
          },
          "403": {
            "description": "Sin permisos"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear un nuevo plan",
        "tags": ["Plans"]
      },
      "get": {
        "operationId": "PlansController_findAllPlans",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de planes"
          },
          "401": {
            "description": "Unauthorized — JWT inválido o ausente"
          },
          "403": {
            "description": "Sin permisos"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar todos los planes",
        "tags": ["Plans"]
      }
    },
    "/api/v1/plans/{id}": {
      "get": {
        "operationId": "PlansController_findOnePlan",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan encontrado"
          },
          "401": {
            "description": "Unauthorized — JWT inválido o ausente"
          },
          "403": {
            "description": "Sin permisos"
          },
          "404": {
            "description": "Plan no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener un plan por ID",
        "tags": ["Plans"]
      },
      "patch": {
        "operationId": "PlansController_updatePlan",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanDto"
              },
              "examples": {
                "update": {
                  "summary": "Actualizar nombre y precio",
                  "value": {
                    "name": "Starter Plus",
                    "price": 24.99
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan actualizado"
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "description": "Unauthorized — JWT inválido o ausente"
          },
          "403": {
            "description": "Sin permisos"
          },
          "404": {
            "description": "Plan no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar un plan por ID",
        "tags": ["Plans"]
      },
      "delete": {
        "operationId": "PlansController_removePlan",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan eliminado"
          },
          "401": {
            "description": "Unauthorized — JWT inválido o ausente"
          },
          "403": {
            "description": "Sin permisos"
          },
          "404": {
            "description": "Plan no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar un plan por ID",
        "tags": ["Plans"]
      }
    },
...
  },
  "info": {
    "title": "API Logistics",
    "description": "Plataforma de gestión logística multi-tenant.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Ingresa tu JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreatePermissionDefinitionDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePermissionDefinitionDto": {
        "type": "object",
        "properties": {}
      },
      "AssignPermissionDto": {
        "type": "object",
        "properties": {}
      },
      "CreatePlanDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePlanDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}